convergency and verification

+++ main.cpp
@@ -4,25 +4,25 @@
 #include<cstdlib>
 #include<vector>
 
-//const float lambda=250e-6;
-//const float channel_width=1e-3;
+//const double lambda=250e-6;
+//const double channel_width=1e-3;
 //const int steps_channel = 1e6 +1; //uneven to hit 0 axis
-//const float d_emitter=1e-2;
-//const float d_sensor=1e-2;
+//const double d_emitter=1e-2;
+//const double d_sensor=1e-2;
 //const int steps_sensor=100;
-//const float x_max=15e-3;
+//const double x_max=15e-3;
 
 
 using namespace std;
 
 
-void print_results(vector<float> result, int x_max, int steps_sensor){
+void print_results(vector<double> result, int x_max, int steps_sensor){
   for (int i = 0; i < (int)result.size(); ++i){
-    cout <<(float)i/(steps_sensor-1) * x_max << "\t" << result[i] << endl;
+    cout <<(double)i/(steps_sensor-1) * x_max << "\t" << result[i] << endl;
   }
 }
 
-void write_results(vector<float> results, char** argv){
+void write_results(vector<double> results, char** argv){
   ofstream out(argv[8]);
   if(out.fail()){
     cout << "outfile cannot be created" << endl;
@@ -33,26 +33,26 @@ void write_results(vector<float> results, char** argv){
 
   //write data
   for (int i = 0; i < (int)results.size(); ++i){
-    out << ( (float)i/(atof(argv[4])-1)*atof(argv[7]) ) << "\t" << results[i] << endl;
+    out << ( (double)i/(atof(argv[4])-1)*atof(argv[7]) ) << "\t" << results[i] << endl;
   }
 
 
 }
 
 
-vector<float> calculate(float lambda, float channel_width, int steps_channel, int steps_sensor, float d_emitter, float d_sensor, float x_max ){
-  vector<float> results(steps_sensor);
+vector<double> calculate(double lambda, double channel_width, int steps_channel, int steps_sensor, double d_emitter, double d_sensor, double x_max ){
+  vector<double> results(steps_sensor);
   //every position on sensor
   #pragma omp parallel for 
   for (int i = 0; i < steps_sensor; ++i){
-      float x = (float)i/(steps_sensor-1) * x_max;
-      float efield_real = 0;
-      float efield_imag = 0;
+      double x = (double)i/(steps_sensor-1) * x_max;
+      double efield_real = 0;
+      double efield_imag = 0;
     //every position in channel (aperture)
     for (int j = 0;  j < steps_channel; j++){
-      float pos_channel = (float)j/(steps_channel-1)*channel_width - channel_width/2;
-      float distance = sqrt( pow(d_emitter,2) + pow(pos_channel,2)  ) + sqrt( pow(d_sensor,2) + pow(x-pos_channel,2)  );
-      float phase = 2.*M_PI / lambda * distance;
+      double pos_channel = (double)j/(steps_channel-1)*channel_width - channel_width/2;
+      double distance =  sqrt( pow(d_emitter,2) + pow(pos_channel,2)  ) + sqrt( pow(d_sensor,2) + pow(x-pos_channel,2)  );
+      double phase = 2.*M_PI / lambda * distance;
       efield_real += cos(phase) / steps_channel;
       efield_imag += sin(phase) / steps_channel;
      }
@@ -71,7 +71,7 @@ int main(int argc, char** argv){
     } 
     
 
-    vector<float> results = calculate(atof(argv[1]), atof(argv[2]), atoi(argv[3]), atoi(argv[4]), atof(argv[5]), atof(argv[6]), atof(argv[7]));
+    vector<double> results = calculate(atof(argv[1]), atof(argv[2]), atoi(argv[3]), atoi(argv[4]), atof(argv[5]), atof(argv[6]), atof(argv[7]));
     
     //print_results(results, atoi(argv[7]), atof(argv[4]));
